Revert rename scroll-to
